% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_sine.R, R/double_triangle.R
\name{double_triangle}
\alias{double_triangle}
\title{Compute the degree day from temperature data.}
\source{
\url{http://biomet.ucdavis.edu/DegreeDays/DegDay.htm}

\url{http://biomet.ucdavis.edu/DegreeDays/DegDay.htm}
}
\usage{
double_triangle(min_temp, max_temp, min_temp_lag, t_L, t_U)

double_triangle(min_temp, max_temp, min_temp_lag, t_L, t_U)
}
\arguments{
\item{min_temp}{Minimum daily temperature}

\item{max_temp}{Maximum daily temperature}

\item{min_temp_lag}{The next day's minimum daily temperature}

\item{t_L}{Lower bound on the temperature}

\item{t_U}{Upper bound on the temperature}

\item{min_temp}{Minimum daily temperature}

\item{max_temp}{Maximum daily temperature}

\item{min_temp_lag}{The next day's minimum daily temperature}

\item{t_L}{Lower bound on the temperature}

\item{t_U}{Upper bound on the temperature}
}
\value{
Degree day from the list of numbers.

Degree day from the list of numbers.
}
\description{
\code{double_sine} Computes the degree days from temperature data, using interpolation from min and max temperature.

\code{double_triangle} Computes the degree days from temperature data, using interpolation from min and max temperature.
}
\examples{

double_sine(10,15,12,11,16)

double_triangle(10,15,12,11,16)
}
\seealso{
For the equations to compute the degree day: \url{http://ipm.ucanr.edu/WEATHER/ddds_tbl.html}.  Because this is done over a 12 day period, we have an extra minimum temperature to compute, hence the extra minimum temperature

For the equations to compute the degree day: \url{http://ipm.ucanr.edu/WEATHER/dddt_tbl.html}.  Because this is done over a 12 day period, we have an extra minimum temperature to compute
}
