% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exuviae.R
\docType{data}
\name{exuviae}
\alias{exuviae}
\title{Exuviae Rate Data}
\format{A data frame with 80 rows and 2 variables}
\source{
Provided by Maia Crews-Erjavec
}
\usage{
data(exuviae)
}
\description{
\itemize{
  \item time. Day measurement taken (a POSIX ct date)
  \item rate. A. canadensis exuviae found per minute
}
}
\keyword{datasets}
