% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_lake.R
\docType{data}
\name{forest_lake}
\alias{forest_lake}
\title{Measured minimum and maximum daily temperature from Forest Lake Weather Station}
\format{A data frame with 365 rows and 4 variables}
\source{
This data has been provided by the Utah Climate Center.

Station ID	USC00212881
Station Name	FOREST LAKE 5NE
Latitude	45.3397
Longitude	-92.9125
Elevation	280.4 m
}
\usage{
data(forest_lake)
}
\description{
\itemize{
  \item year. Year of measurement
  \item day. Day of the year
  \item min_temp. Minimum temperature (degrees Celsius)
  \item max_temp. Maximum temperature (degrees Celsius)
}
}
\keyword{datasets}
