% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/railtracks.R
\docType{data}
\name{railtracks}
\alias{railtracks}
\title{Measured minimum and maximum daily temperature from Railtracks Pond}
\format{A data frame with 168 rows and 4 variables}
\source{
Provided by Emily Schilling
}
\usage{
data(railtracks)
}
\description{
\itemize{
  \item year. Year of measurement
  \item day. Day of the year
  \item min_temp. Minimum temperature (degrees Celsius)
  \item max_temp. Maximum temperature (degrees Celsius)
}
}
\keyword{datasets}
