% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sine.R
\name{single_sine}
\alias{single_sine}
\title{Compute the degree day from temperature data.}
\source{
\url{http://biomet.ucdavis.edu/DegreeDays/DegDay.htm}
}
\usage{
single_sine(min_temp, max_temp, t_L, t_U)
}
\arguments{
\item{min_temp}{Minimum daily temperature}

\item{max_temp}{Maximum daily temperature}

\item{t_L}{Lower bound on the temperature}

\item{t_U}{Upper bound on the temperature}
}
\value{
Degree day from the list of numbers.
}
\description{
\code{single_sine} Computes the degree days from temperature data, using interpolation from min and max temperature.
}
\examples{

single_sine(10,15,11,16)
}
\seealso{
For the equations to compute the degree day: \url{http://ipm.ucanr.edu/WEATHER/ddss_tbl.html}
}
