% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_triangle.R
\name{single_triangle}
\alias{single_triangle}
\title{Compute the degree day from temperature data.}
\usage{
single_triangle(min_t, max_t, t_L, t_U)
}
\arguments{
\item{min_t}{Minimum daily temperature}

\item{max_t}{Maximum daily temperature}

\item{t_L}{Lower bound on the temperature}

\item{t_U}{Upper bound on the temperature}
}
\value{
Degree day from the list of numbers.
}
\description{
\code{single_triangle} Computes the degree days from temperature data, using interpolation from
}
\examples{

single_triangle(10,15,11,16)
}
\seealso{
For the equations to compute the degree day: \url{http://ipm.ucanr.edu/WEATHER/ddst_tbl.html}
}
